Convert a simulations_collection to a tibble

Cumulative statistics are shown to gauge how the simulations converge.

# S3 method for simulations_collection
as_tibble(x, target_dose = NULL, alpha = 0.05, ...)

Arguments

x

object of type simulations_collection

target_dose

numerical dose index, or NULL (default) for all doses

alpha

significance level for symmetrical confidence intervals

...

extra args are ignored

Value

a tibble with cols:

"dose"the dose-level

"n"cumulative inference using the first n simulated iterations

"design.x"The first design in the comparison, aka design X

"hit.x"logical showing if design X recommended dose in iterate n

"design.y"The second design in the comparison, aka design Y

"hit.x"logical showing if design Y recommended dose in iterate n

"X"cumulative sum of hit.x within dose, i.e. count of recommendations

"X2"cumulative sum of hit.x^2 within dose

"Y"cumulative sum of hit.y within dose, i.e. count of recommendations

"Y2"cumulative sum of hit.y^2 within dose

"XY"cumulative sum of hit.x * hit.y within dose

"psi1"X / n

"psi2"Y / n

"v_psi1"variance of psi1

"v_psi2"variance of psi2

"cov_psi12"covariance of psi1 and psi2

"delta"psi1 - psi2

"v_delta"variance of delta

"se_delta"standard error of delta

"delta_l"delta - q * se_delta, where q is alpha / 2 normal quantile

"delta_u"delta + q * se_delta, where q is alpha / 2 normal quantile

"comparison"Label of design.x vs design.y, using design names