Expand the cohort of the last given dose to at least n patients

Expand the cohort of the last given dose to at least n patients

expand_last_dose_to_cohort(parent_selector_factory, n)

Arguments

parent_selector_factory

Object of type selector_factory.

n

Continue at current dose until at least n patients are evaluated.

Value

an object of type selector_factory that can fit a dose-finding model to outcomes.

Examples

model <- follow_path(path = "1N 2N 3N") %>%
  expand_last_dose_to_cohort(n = 3)
x <- model %>% fit("1N 2N 3T")
# Design will continue:
continue(x)
#> [1] TRUE
# at dose 3:
recommended_dose(x)
#> [1] 3
# because it is expanding the last dose to a cohort of three.