Get an object to fit the BOIN COMB model using the BOIN package.

Get an object to fit the BOIN COMB model using the BOIN package.

Usage

get_boin_comb(
  parent_selector_factory = NULL,
  num_doses,
  target,
  use_stopping_rule = TRUE,
  ...
)

Arguments

parent_selector_factory

optional object of type selector_factory that is in charge of dose selection before this class gets involved. Leave as NULL to just use CRM from the start.

num_doses

integer vector of the number of doses of treatment 1, 2

target

We seek a dose with this probability of toxicity.

use_stopping_rule

TODO

...

Extra args are passed to next.comb.

Value

an object of type selector_factory that can fit the BOIN COMB model to outcomes.

References

Lin, R., & Yin, G. (2017). Bayesian optimal interval design for dose finding in drug-combination trials. Statistical methods in medical research, 26(5), 2155-2167.

Examples

num_doses <- c(3, 4)
target <- 0.25
boin_fitter <- get_boin_comb(num_doses = num_doses, target = target)
x1 <- fit(boin_fitter, outcomes = "1.1NNN")
x1
#> Patient-level data:
#> # A tibble: 3 × 5
#>   Cohort Patient Dose_string Dose        Tox
#>    <int>   <int> <chr>       <list>    <int>
#> 1      1       1 1.1         <int [2]>     0
#> 2      1       2 1.1         <int [2]>     0
#> 3      1       3 1.1         <int [2]>     0
#> 
#> Dose-level data:
#> # A tibble: 13 × 8
#>    dose     tox     n empiric_tox_rate mean_prob_tox median_prob_tox admissible
#>    <ord>  <dbl> <dbl>            <dbl>         <dbl>           <dbl> <lgl>     
#>  1 NoDose     0     0                0        0          0           TRUE      
#>  2 1.1        0     3                0        0.0161     0.000000215 TRUE      
#>  3 1.2        0     0              NaN        0.5        0.500       TRUE      
#>  4 1.3        0     0              NaN        0.5        0.500       TRUE      
#>  5 1.4        0     0              NaN        0.5        0.500       TRUE      
#>  6 2.1        0     0              NaN        0.5        0.500       TRUE      
#>  7 2.2        0     0              NaN        0.5        0.500       TRUE      
#>  8 2.3        0     0              NaN        0.5        0.500       TRUE      
#>  9 2.4        0     0              NaN        0.5        0.500       TRUE      
#> 10 3.1        0     0              NaN        0.5        0.500       TRUE      
#> 11 3.2        0     0              NaN        0.5        0.500       TRUE      
#> 12 3.3        0     0              NaN        0.5        0.500       TRUE      
#> 13 3.4        0     0              NaN        0.5        0.500       TRUE      
#> # ℹ 1 more variable: recommended <lgl>
#> 
#> The model targets a toxicity level of 0.25.
#> The model advocates continuing at dose 1.2.
x2 <- fit(boin_fitter, outcomes = "1.1NNN 2.1TNT")
x2
#> Patient-level data:
#> # A tibble: 6 × 5
#>   Cohort Patient Dose_string Dose        Tox
#>    <int>   <int> <chr>       <list>    <int>
#> 1      1       1 1.1         <int [2]>     0
#> 2      1       2 1.1         <int [2]>     0
#> 3      1       3 1.1         <int [2]>     0
#> 4      2       4 2.1         <int [2]>     1
#> 5      2       5 2.1         <int [2]>     0
#> 6      2       6 2.1         <int [2]>     1
#> 
#> Dose-level data:
#> # A tibble: 13 × 8
#>    dose     tox     n empiric_tox_rate mean_prob_tox median_prob_tox admissible
#>    <ord>  <dbl> <dbl>            <dbl>         <dbl>           <dbl> <lgl>     
#>  1 NoDose     0     0            0            0          0           TRUE      
#>  2 1.1        0     3            0            0.0161     0.000000215 TRUE      
#>  3 1.2        0     0          NaN            0.5        0.500       TRUE      
#>  4 1.3        0     0          NaN            0.5        0.500       TRUE      
#>  5 1.4        0     0          NaN            0.5        0.500       TRUE      
#>  6 2.1        2     3            0.667        0.632      0.525       TRUE      
#>  7 2.2        0     0          NaN            0.632      0.525       TRUE      
#>  8 2.3        0     0          NaN            0.632      0.525       TRUE      
#>  9 2.4        0     0          NaN            0.632      0.525       TRUE      
#> 10 3.1        0     0          NaN            0.632      0.525       TRUE      
#> 11 3.2        0     0          NaN            0.632      0.525       TRUE      
#> 12 3.3        0     0          NaN            0.632      0.525       TRUE      
#> 13 3.4        0     0          NaN            0.632      0.525       TRUE      
#> # ℹ 1 more variable: recommended <lgl>
#> 
#> The model targets a toxicity level of 0.25.
#> The model advocates continuing at dose 1.1.