Get an object to fit the TITE-CRM model using the dfcrm package.

Get an object to fit the TITE-CRM model using the dfcrm package.

Usage

get_dfcrm_tite(parent_selector_factory = NULL, skeleton, target, ...)

Arguments

parent_selector_factory

optional object of type selector_factory that is in charge of dose selection before this class gets involved. Leave as NULL to just use CRM from the start.

skeleton

Dose-toxicity skeleton, a non-decreasing vector of probabilities.

target

We seek a dose with this probability of toxicity.

...

Extra args are passed to crm.

Value

an object of type selector_factory that can fit the CRM model to outcomes.

Details

This function is a short-cut to get_dfcrm(tite = TRUE). See get_dfcrm for full details.

References

Cheung, K. 2019. dfcrm: Dose-Finding by the Continual Reassessment Method. R package version 0.2-2.1. https://CRAN.R-project.org/package=dfcrm

Cheung, K. 2011. Dose Finding by the Continual Reassessment Method. Chapman and Hall/CRC. ISBN 9781420091519

Examples

skeleton <- c(0.05, 0.1, 0.25, 0.4, 0.6)
target <- 0.25
model1 <- get_dfcrm_tite(skeleton = skeleton, target = target)
outcomes <- data.frame(
  dose = c(1, 1, 2, 2, 3, 3),
  tox = c(0, 0, 0, 0, 1, 0),
  weight = c(1, 1, 1, 0.9, 1, 0.5),
  cohort = c(1, 2, 3, 4, 5, 6)
)
fit <- model1 %>% fit(outcomes)