Get an object to fit the mTPI dose-finding model.

The modified toxicity probability interval (mTPI)is a dose-escalation design by Ji et al. As the name suggests, it is an adaptation of the TPI design.

get_mtpi(
  num_doses,
  target,
  epsilon1,
  epsilon2,
  exclusion_certainty,
  alpha = 1,
  beta = 1,
  ...
)

Arguments

num_doses
Number of doses under investigation.

target
We seek a dose with this probability of toxicity.

epsilon1
This parameter determines the lower bound of the equivalence interval. See Details.

epsilon2
This parameter determines the upper bound of the equivalence interval. See Details.

exclusion_certainty
Numeric, threshold posterior certainty required to exclude a dose for being excessively toxic. The authors discuss values in the range 0.7 - 0.95. Set to a value > 1 to suppress the dose exclusion mechanism. The authors use the Greek letter xi for this parameter.

alpha
First shape parameter of the beta prior distribution on the probability of toxicity.

beta
Second shape parameter of the beta prior distribution on the probability of toxicity.

...
Extra args are passed onwards.

Value

an object of type selector_factory that can fit the TPI model to outcomes.

Details

The design seeks a dose with probability of toxicity \(p_{i}\) close to a target probability \(p_{T}\) by iteratively calculating the interval $$p_{T} - \epsilon_{1} < p_{i} < p_{T} + \epsilon_{2}$$ In this model, \(\epsilon_{1}\) and \(\epsilon_{2}\) are specified constants. \(p_{i}\) is estimated by a Bayesian beta-binomial conjugate model $$p_{i} | data \sim Beta(\alpha + x_{1}, \beta + n_{i} - x_{i}),$$ where \(x_{i}\) is the number of toxicities observed and \(n_{i}\) is the number of patients treated at dose \(i\), and \(\alpha\) and \(\beta\) are hyperparameters for the beta prior on \(p_{i}\). A dose is excluded as inadmissible if $$P(p_{i} > p_{T} | data) > \xi$$ The trial commences at a starting dose, possibly dose 1. If dose \(i\) has just been evaluated in patient(s), dose selection decisions proceed by calculating the unit probability mass of the true toxicity rate at dose \(i\) using the partition of the probability space \(p_{i} < p_{T} - \epsilon_{1}\), \(p_{T} - \epsilon_{1} < p_{i} < p_{T} + \epsilon_{2}\), and \(p_{i} > p_{T} + \epsilon_{2}\). The unit probability mass (UPM) of an interval is the posterior probability that the true toxicity rate belongs to the interval divided by the width of the interval. The interval with maximal UPM determines the recommendation for the next patient(s), with the intervals corresponding to decisions tp escalate, stay, and de-escalate dose, respectively. Further to this are rules that prevent escalation to an inadmissible dose. In their paper, the authors demonstrate acceptable operating performance using \(\alpha = \beta = 1\), \(K_{1} = 1\), \(K_{2} = 1.5\) and \(\xi = 0.95\). See the publications for full details.

References

Ji, Y., Liu, P., Li, Y., & Bekele, B. N. (2010). A modified toxicity probability interval method for dose-finding trials. Clinical Trials, 7(6), 653–663. https://doi.org/10.1177/1740774510382799

Ji, Y., & Yang, S. (2017). On the Interval-Based Dose-Finding Designs, 1–26. Retrieved from http://arxiv.org/abs/1706.03277

Examples

target <- 0.25
model1 <- get_mtpi(num_doses = 5, target = target, epsilon1 = 0.05,
  epsilon2 = 0.05, exclusion_certainty = 0.95)

outcomes <- '1NNN 2NTN'
model1 %>% fit(outcomes) %>% recommended_dose()

#> [1] 2