Visualise dose-paths as a graph

Visualise dose-paths as a graph

graph_paths(paths, viridis_palette = "viridis", RColorBrewer_palette = NULL)

Arguments

paths

Object of type dose_paths

viridis_palette

optional name of a colour palette in the viridis package.

RColorBrewer_palette

optional name of a colour palette in the RColorBrewer package.

Details

The viridis package supports palettes: viridis, magma, plasma, inferno, and cividis. The RColorBrewer package supports many palettes. Refer to those packages on CRAN for more details.

Examples

paths <- get_three_plus_three(num_doses = 5) %>%
  get_dose_paths(cohort_sizes = c(3, 3, 3))
if (FALSE) { # \dontrun{
graph_paths(paths)
graph_paths(paths, viridis_palette = 'plasma')
graph_paths(paths, RColorBrewer_palette = 'YlOrRd')
} # }