Select dose by the CIBP selection criterion.

This method selects dose by the convex infinite bounds penalisation (CIBP) criterion of Mozgunov & Jaki. Their method is mindful of the uncertainty in the estimates of the probability of toxicity and uses an asymmetry parameter to penalise escalation to risky doses.

select_dose_by_cibp(parent_selector_factory, a, target = NULL)

Arguments

parent_selector_factory

Object of type selector_factory.

a

Number between 0 and 2, the asymmetry parameter. See References.

target

We seek a dose with this probability of toxicity. If not provided, the value will be sought from the parent dose-selector.

Value

an object of type selector_factory that can fit a dose-finding model to outcomes.

References

Mozgunov P, Jaki T. Improving safety of the continual reassessment method via a modified allocation rule. Statistics in Medicine.1-17. doi:10.1002/sim.8450

Examples

skeleton <- c(0.05, 0.1, 0.25, 0.4, 0.6)
target <- 0.33

# Let's compare escalation behaviour of a CRM model without CIBP criterion:
model1 <- get_dfcrm(skeleton = skeleton, target = target)
# To one with the CIBP criterion:
model2 <- get_dfcrm(skeleton = skeleton, target = target) %>%
  select_dose_by_cibp(a = 0.3)

# Despite one-in-three tox at first dose, regular model is ready to escalate:
model1 %>% fit('1NTN') %>% recommended_dose()
#> [1] 2
# But the model using CIBP is more risk averse:
model2 %>% fit('1NTN') %>% recommended_dose()
#> [1] 1