Make an instance of type simulations_collection

This object can be cast to a tibble with as_tibble to generate useful pairwise comparisons of the probability of recommending each dose for each pair of designs investigated. See as_tibble.simulations_collection for a description.

Usage

simulations_collection(sim_map)

Arguments

sim_map

list, character -> simulations object

Value

object of class simulations_collection, inheriting from list

References

Sweeting, M., Slade, D., Jackson, D., & Brock, K. (2024). Potential outcome simulation for efficient head-to-head comparison of adaptive dose-finding designs. arXiv preprint arXiv:2402.15460