Get an object to fit the BOIN COMB model using the BOIN package.

Get an object to fit the BOIN COMB model using the BOIN package.

get_boin_comb(num_doses, target, use_stopping_rule = TRUE, ...)

Arguments

num_doses

integer vector of the number of doses of treatment 1, 2

target

We seek a dose with this probability of toxicity.

use_stopping_rule

TODO

...

Extra args are passed to next.comb.

Value

an object of type selector_factory that can fit the BOIN COMB model to outcomes.

References

Lin, R., & Yin, G. (2017). Bayesian optimal interval design for dose finding in drug-combination trials. Statistical methods in medical research, 26(5), 2155-2167.

Examples

# TODO