Simulated trials.

This class encapsulates that many notional or virtual trials can be simulated. Each recommends a dose or doses, keeps track of how many patients have been treated at what doses, what toxicity outcomes have been seen, and whether a trial advocates continuing. We run simulations to learn about the operating characteristics of a trial design.

Computationally, the simulations class supports much of the same interface as selector, and a little more.

simulations()

Details

The simulations object implements the following functions:

num_patients

num_doses

dose_indices

n_at_dose

tox_at_dose

num_tox

recommended_dose

prob_administer

prob_recommend

trial_duration

See also

Examples

# Simulate performance of the 3+3 design:
true_prob_tox <- c(0.12, 0.27, 0.44, 0.53, 0.57)
sims <- get_three_plus_three(num_doses = 5) %>%
  simulate_trials(num_sims = 50, true_prob_tox = true_prob_tox)
# The returned object has type 'simulations'. The supported interface is:
sims %>% num_patients()

#>  [1]  9  9 12 12 12 12  6  3  9 15  9 12  9 12  6 18  9  6  6  9  6 12 18 15  9
#> [26] 15  9 15 12 18 12  6  9  6 18 12  9  9 15 12 12  9  6  9 12  6 12 24 18 12

sims %>% num_doses()

#> [1] 5

sims %>% dose_indices()

#> [1] 1 2 3 4 5

sims %>% n_at_dose()

#> # A tibble: 50 x 5
#>      `1`   `2`   `3`   `4`   `5`
#>    <int> <int> <int> <int> <int>
#>  1     3     3     3     0     0
#>  2     3     6     0     0     0
#>  3     3     6     3     0     0
#>  4     3     6     3     0     0
#>  5     3     3     6     0     0
#>  6     3     6     3     0     0
#>  7     6     0     0     0     0
#>  8     3     0     0     0     0
#>  9     3     3     3     0     0
#> 10     3     3     6     3     0
#> # … with 40 more rows

sims %>% tox_at_dose()

#> # A tibble: 50 x 5
#>      `1`   `2`   `3`   `4`   `5`
#>    <int> <int> <int> <int> <int>
#>  1     0     0     2     0     0
#>  2     0     2     0     0     0
#>  3     0     1     2     0     0
#>  4     0     1     2     0     0
#>  5     0     0     2     0     0
#>  6     0     1     3     0     0
#>  7     2     0     0     0     0
#>  8     2     0     0     0     0
#>  9     0     0     2     0     0
#> 10     0     0     1     2     0
#> # … with 40 more rows

sims %>% num_tox()

#>  [1] 2 2 3 3 2 4 2 2 2 3 2 3 3 2 2 5 2 3 2 2 2 4 3 3 3 3 2 3 3 5 3 2 3 2 3 2 3 2
#> [39] 3 2 3 2 2 2 3 2 3 5 4 4

sims %>% recommended_dose()

#>  [1]  2  1  2  2  2  2 NA NA  2  3  1  2  1  3 NA  3  2 NA NA  1 NA  1  3  2  1
#> [26]  3  2  3  2  2  2 NA  1 NA  3  3  1  2  3  2  1  1 NA  1  2  1  2  4  3  2

sims %>% prob_administer()

#>           1           2           3           4           5 
#> 0.380434783 0.320652174 0.211956522 0.081521739 0.005434783 

sims %>% prob_recommend()

#> NoDose      1      2      3      4      5 
#>   0.18   0.24   0.36   0.20   0.02   0.00 

sims %>% trial_duration()

#>  [1] 11.4  5.5 16.8 11.1 14.5 12.6  8.3  2.7  6.8 15.2  6.8  9.2  5.0  7.1  5.3
#> [16] 22.6  9.4  5.6  9.2  9.3  4.5 12.3 16.9 13.5 18.0  7.3  9.2 12.4  9.2 15.6
#> [31]  7.9  6.8  5.3  6.2 19.8 13.7  8.7  6.8 15.6 15.1  6.5 10.5  5.9  7.4 14.9
#> [46]  6.4  5.3 30.6 14.6  7.4